Missing serialization methods are now reported during compile time.
Instead of using a SFINAE implementation to detect whether a
serialization member method exists for a given type, the user of
the library now has to register types without serialization methods
through moose::RegisterEmptyType instead.

This allows for compile time errors and simplifies the implementation.

This also fixes Linux compilation (tested with GCC 7.4.0).