Type: Use `type_index` instead of hash. No longer rely on names.
Classes and base classes are now identified by their type_index. This
should allow for much faster class hierarchy checks. Furthermore, it
is no longer necessary to obtain the types of base classes during
registration of a new type. This allows for registering classes in
arbitrary order, which is important for registration during static
initialization.